Separate liveness and readiness checks
